fix: SpaceResponse docs
